requirements.txt
django==1.9.7
djangorestframework
django-ckeditor
pillow

